Draw bar chart with a single datapoint.
This patch allows rendering of bar chart that only has a single
datapoint. If a scaleStartValue is not defined the minimum value is set
to 0.

If a scaleStartValue is defined it is used as the minimum value in the
chart.

Without this patch a chart with a single datapoint is rendered with a
bar height of zero, which makes the chart appear to be empty.