Fixing legend fonts after tooltip display
When a Tooltip or MultiTooltip is rendered, the canvas context font is set
to the Tooltip Style. After te font has been rendered, the font is never
set back to what it had been before. Causing the legend to be rendered in
the same font.

Made changes to keep track of the previous font before rendering then
restoring it afterwards.