Option to only display points within cursor area
By setting "tooltipCursorSensitive" to true, only the data points within
"tooltipCursorHitPixels" will be displayed in the tooltip box. Except
for data points with a "custom value" (those not numeric, not null),
those will display anywhere on the y-axis, same as current default.
Also, a minor fix to account for the height of the blurry pixels (when
using shadow)  when determining that the tooltip box is bleeding outside
of the canvas.