Initial work for enabling arrays for bar styling
I added a function to Chart.helpers called nthOrIdentity which takes in
two arguments, a potential array called 'potentialArray' and a desired
index 'nthIndex'.  If 'potentialArray' is an array it will return the item
at 'nthIndex', else it will simply return 'potentialArray'.

I have also updated Chart.Bar.js to use this helper function when
setting the styles for the bars.