Introduced helper function isArray
This is a mild shim around Array.isArray so one doesn't need to test for
the existence of Array.isArray -- which might not exist in certain
browsers, for example IE7 and IE8 -- while still being able to use it if
it does exist.

I have also converted nthOrIdentity to use this new function.