Created more concise calls to nthOrIdentity
Chart.Bar.js used to have two calls to helpers.nthOrIdentity for the
highlight variables which, if the first is passed an undefined object (which
it might be), could mean that all of nthOrIdentity would be called only
to return 'undefined', then called again with defined value.

To avoid this, I moved the variable comparison inside the call to
helpers.nthOrIdentity, so the function will only be called once.