transitions on charts that start with no values calculate based on NaN and cannot take live updates
Initial calculations on, for example, a Pie or Doughnut chart rely on the chart having an initial 'total' to
calculate circumference et. al.  When this value is initially zero (a chart with no data) then the segments
cannot transition out of this state due to the "ease" calculation.  This patch applies the value directly and
skips easing when the existing segment value is NaN.