Added original tooltip anchor position to the model
When using a custom tooltip, the position returned on the tooltip model contains the position adjusted for the default tooltip of the library. This is very cumbersome to use when the custom tooltip does not conform to the same dimensions or page constraints as the default tooltip. Adding these anchor coordinates allow the developer to orient themselves to the original position of the data point on the graph.