legendCallback improvement
Since fixing the legendCallback before, i've noticed that the legend produced by it looks nothing like the one produced on the canvas. So i've improved the styling to make it look as close to the canvas one as possible. The font of the generated Legend can now also be set in the chart's options i.e. Options: { display: false, legend: { fontFamily: "insert font family here" } }, this will stop the Canvas legend from displaying but still set the font for the generated legend. then the user must call generateLegend() and apply the generated html to the inner html of an element as per usual.