Updated legendCallback
Have improved the legendCallback function so that the legend will look like the one generated on the canvas. The user can now set the font for the legend in the chart's options like you would usually, but they will need to set display to false for the legend and instead call the generateLegend function if they want the HTML version.