Fix wild special zero grid line when undefined
Remove the special drawing for an undefined zero grid line since it causes issue when the first tick is not anymore aligned on the scale extremity. Only linear based scales now display a special grid line for the origin. Hide scales in the filler plugin unit test fixtures to avoid future failures due to changes unrelated to the tested features.