CHanges as requested simonbrunel
A few early feedback:

instead of duplicating the logic to pick the "right" value in each scale, it may be better to move it directly in getRightValue()? Though, that would need some investigation to be sure it doesn't break anything.
I would not introduce the new borderSkipped: 'none' value, but instead support borderSkipped: false|null.
I would not call it lowY or highY but low/hight or min/max