CHanges to float-bar support.
parseValue is in scale.js, returns data as object with min, max and val. This functions is used in controller.bar.js to calculate start point and gap for floating-bar. Also this functiotn is used in scale.linearjs and scale.logarithmic.js to calculate scale min and max values and label texts.