Use MutationObserver to detect when canvas is attached
Using CSS animations to detect when the canvas is attached causes issues when dealing with CSP. Instead, we can use the MutationObserver API to efficiently detect when a node is added or removed, which is now widely adopted (https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver).

MutationObserver is not supported on IE10, however we use this feature only to detect when a node is attached to the DOM. Checking if MutationObserver is defined allows us to keep support for old browsers when the node is already attached, which is the case in most situations.