Docs: Setting maxTicksLimit disabled autoSkip
Spent quite a long time debugging why maxTicksLimit wasn't working with autoSkip. I expected autoSkip with maxTicksLimit to use the autoSkip number of ticks if it was smaller or maxTicksLimit number of ticks if it was smaller. There doesn't seem to be a way to do this at the moment. I could do a PR for a new autoSkipMaxTicks if there was interest.

The code that finally gave me my aha moment:
function autoSkip(scale, ticks) {
...
    const ticksLimit = tickOpts.maxTicksLimit || determineMaxTicks(scale);