docs:  explain how to build cdn js file
I was trying to figure out how to re-build the same file that is served from CDNs, and I got confused and tripped over myself on the way.  through my own mistakes, I had actually convinced myself that the build script in this repo did not produce the file the the CDN was serving.   Nothing in the docs talked about it, so I just assumed it was some "special thing that exists somewhere else."

I am adding this to the docs now so that the next person reading through will be re-assured and hopefully find the right file.